Remove "non inline" images option
It's been more than two years since I touched this code, so I don't
remember what I was thinking, but using non inline images is pointless for
this tool. Non inline images are better when you can use one image
multiple times in the same PDF, but that simply doesn't happen in this
tool, ever, so there is no size advantage.

As for inline images being problematic, it's not a problem with inlining
them, but with poorly implemented viewers. Most viewers don't have a
problem with this, and since it's the viewer fault, whoever finds trouble
with this should fix it or look for a better viewer.