Remove unnecesary functions
PDF text compression is only of any use when text is stored in the PDF
file. Since this tool only stores images, this option is useless.

Rotation is out of the scope of this tool. If the images need any kind of
processing, this is the wrong tool to do that. K.I.S.S.