Remove resizing option
Removal of another image processing option. There are a lot of tools to
manipulate images out there. If they must be of an specific size, proccess
them beforehand.

This fixes a bug in the image manipulation where the script was outputing
bitmaps to the PDF instead of JPEG files, producing REALLY BIG files.

A quick test produces a 1,5MB PDF where before this change would produce
a 22MB one, showing that K.I.S.S. is a good principle.